requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
google-auth==2.40.2
google-genai==1.18.0
googlesearch-python==1.3.0
greenlet==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-whois==0.9.5
pytz==2025.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.1
threadpoolctl==3.6.0
tld==0.13
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
websockets==15.0.1
whois==1.20240129.2
